config.py
import os
from dotenv import load_dotenv, find_dotenv


load_dotenv(find_dotenv())


class Config:
    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    VERBOSE = bool(os.environ.get('VERBOSE_FLASK'))
